requirements.txt
# Core dependencies
watchdog>=3.0.0          # Real-time folder monitoring
pdfplumber>=0.10.0       # PDF text extraction for smart categorization
python-docx>=1.1.0       # DOCX text extraction
PyYAML>=6.0              # Config file parsing
plyer>=2.1.0             # Cross-platform desktop notifications

# Testing
pytest>=7.0
pytest-cov>=4.0
